Added files via upload
In the order of files created:
Brainstorming.txt is the initial idea. 
battle_demo.py is the very first trial run of a battle simulation. The final code will be different. This is just an idea to give me a direction to work towards.
Layout Draft.txt is my first organization of the functions I will be need as well as other info.
new_weapon.py is a standalone script that will generate a new weapon based on the player level input. When incorporated into the full program, the player level will be extrapolated from a different section of code.